requirement.txt
asgiref==3.3.4
Django==3.2.3
djangorestframework==3.12.4
Faker==8.9.0
mysqlclient==2.0.3
Pillow==8.2.0
psycopg2-binary==2.8.6
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
